Put new integration test into its own class
